app, libgimp, pdb: add g_return*_if_fail() test for array size args.
Generated libgimp functions' arguments are normally tested through the
PDB, so that you get proper error messages when trying to call a
function with invalid arguments.
This is not true anymore for array arguments since the size argument is
not a PDB arg, only in the C function. Therefore I'm adding an
infrastructure asserting on invalid size, using the same PDB type
annotations as other args. It is important to assert valid input on
plug-in side (i.e. libgimp) so that the core doesn't make any assumption
on having received valid input when it has not.

Also changing size argument of gimppainttools PDB-generated functions to
proper gsize.

+++ app/paint/gimppaintcore-stroke.c
@@ -51,7 +51,7 @@ gimp_paint_core_stroke (GimpPaintCore     *core,
                         GimpDrawable      *drawable,
                         GimpPaintOptions  *paint_options,
                         GimpCoords        *strokes,
-                        gint               n_strokes,
+                        gsize              n_strokes,
                         gboolean           push_undo,
                         GError           **error)
 {
