+++ app/paint/gimppaintcore-stroke.h
@@ -23,7 +23,7 @@ gboolean   gimp_paint_core_stroke          (GimpPaintCore      *core,
                                             GimpDrawable       *drawable,
                                             GimpPaintOptions   *paint_options,
                                             GimpCoords         *strokes,
-                                            gint                n_strokes,
+                                            gsize               n_strokes,
                                             gboolean            push_undo,
                                             GError            **error);
 gboolean   gimp_paint_core_stroke_boundary (GimpPaintCore      *core,
