+++ libgimp/gimppainttools_pdb.h
@@ -34,73 +34,73 @@ G_BEGIN_DECLS
 
 gboolean gimp_airbrush           (GimpDrawable             *drawable,
                                   gdouble                   pressure,
-                                  gint                      num_strokes,
+                                  gsize                     num_strokes,
                                   const gdouble            *strokes);
 gboolean gimp_airbrush_default   (GimpDrawable             *drawable,
-                                  gint                      num_strokes,
+                                  gsize                     num_strokes,
                                   const gdouble            *strokes);
 gboolean gimp_clone              (GimpDrawable             *drawable,
                                   GimpDrawable             *src_drawable,
                                   GimpCloneType             clone_type,
                                   gdouble                   src_x,
                                   gdouble                   src_y,
-                                  gint                      num_strokes,
+                                  gsize                     num_strokes,
                                   const gdouble            *strokes);
 gboolean gimp_clone_default      (GimpDrawable             *drawable,
-                                  gint                      num_strokes,
+                                  gsize                     num_strokes,
                                   const gdouble            *strokes);
 gboolean gimp_convolve           (GimpDrawable             *drawable,
                                   gdouble                   pressure,
                                   GimpConvolveType          convolve_type,
-                                  gint                      num_strokes,
+                                  gsize                     num_strokes,
                                   const gdouble            *strokes);
 gboolean gimp_convolve_default   (GimpDrawable             *drawable,
-                                  gint                      num_strokes,
+                                  gsize                     num_strokes,
                                   const gdouble            *strokes);
 gboolean gimp_dodgeburn          (GimpDrawable             *drawable,
                                   gdouble                   exposure,
                                   GimpDodgeBurnType         dodgeburn_type,
                                   GimpTransferMode          dodgeburn_mode,
-                                  gint                      num_strokes,
+                                  gsize                     num_strokes,
                                   const gdouble            *strokes);
 gboolean gimp_dodgeburn_default  (GimpDrawable             *drawable,
-                                  gint                      num_strokes,
+                                  gsize                     num_strokes,
                                   const gdouble            *strokes);
 gboolean gimp_eraser             (GimpDrawable             *drawable,
-                                  gint                      num_strokes,
+                                  gsize                     num_strokes,
                                   const gdouble            *strokes,
                                   GimpBrushApplicationMode  hardness,
                                   GimpPaintApplicationMode  method);
 gboolean gimp_eraser_default     (GimpDrawable             *drawable,
-                                  gint                      num_strokes,
+                                  gsize                     num_strokes,
                                   const gdouble            *strokes);
 gboolean gimp_heal               (GimpDrawable             *drawable,
                                   GimpDrawable             *src_drawable,
                                   gdouble                   src_x,
                                   gdouble                   src_y,
-                                  gint                      num_strokes,
+                                  gsize                     num_strokes,
                                   const gdouble            *strokes);
 gboolean gimp_heal_default       (GimpDrawable             *drawable,
-                                  gint                      num_strokes,
+                                  gsize                     num_strokes,
                                   const gdouble            *strokes);
 gboolean gimp_paintbrush         (GimpDrawable             *drawable,
                                   gdouble                   fade_out,
-                                  gint                      num_strokes,
+                                  gsize                     num_strokes,
                                   const gdouble            *strokes,
                                   GimpPaintApplicationMode  method,
                                   gdouble                   gradient_length);
 gboolean gimp_paintbrush_default (GimpDrawable             *drawable,
-                                  gint                      num_strokes,
+                                  gsize                     num_strokes,
                                   const gdouble            *strokes);
 gboolean gimp_pencil             (GimpDrawable             *drawable,
-                                  gint                      num_strokes,
+                                  gsize                     num_strokes,
                                   const gdouble            *strokes);
 gboolean gimp_smudge             (GimpDrawable             *drawable,
                                   gdouble                   pressure,
-                                  gint                      num_strokes,
+                                  gsize                     num_strokes,
                                   const gdouble            *strokes);
 gboolean gimp_smudge_default     (GimpDrawable             *drawable,
-                                  gint                      num_strokes,
+                                  gsize                     num_strokes,
                                   const gdouble            *strokes);
